requirements.txt
chess
pyserial
kivy[base]
